Add support for ImageSource
