--- package/Codegen/FakeEnum.cs
@@ -1,11 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace Codegen
-{
-    public struct FakeEnum
-    {
-        public string Name { get; set; }
-        public Dictionary<string, int> Values { get; set; }
-    }
-}
-
