+++ package/Codegen/XamlNames.cs
@@ -1,0 +1,10 @@
+﻿namespace Codegen
+{
+    public static class XamlNames
+    {
+        public const string XamlNamespace = "Windows.UI.Xaml";
+        public const string TopLevelProjectedType = $"{XamlNamespace}.DependencyObject";
+        public const string MuxNamespace = "Microsoft.UI.Xaml";
+    }
+}
+
